+++ resources/views/inquiry-product/amounts.blade.php
@@ -411,19 +411,21 @@
         </script>
         <script>
             function deletePartFromAmount(id) {
-                $.ajaxSetup({
-                    headers: {
-                        'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
-                    }
-                });
+                if (confirm('قطعه حذف شود ؟')) {
+                    $.ajaxSetup({
+                        headers: {
+                            'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
+                        }
+                    });
 
-                $.ajax({
-                    type: 'DELETE',
-                    url: '/inquiries/' + id + '/destroy-amount',
-                    success: function () {
-                        location.reload();
-                    }
-                });
+                    $.ajax({
+                        type: 'DELETE',
+                        url: '/inquiries/' + id + '/destroy-amount',
+                        success: function () {
+                            location.reload();
+                        }
+                    });
+                }
             }
         </script>
     </x-slot>
